+++ SourceCode/Engine/Math/Easing.h
@@ -13,4 +13,4 @@ float EaseOutQuint(float start, float end, float movetime, float endtime);
 
 Vector3 EaseOutQuint(Vector3 start, Vector3 end, float movetime, float endtime);
 Vector2 EaseOutQuint(Vector2 start, Vector2 end, float movetime, float endtime);
-double EaseInBack(float t, int start, int end, int num = 1, double p1 = 1.70158);
+float EaseInBack(float t, float start, float end, int num = 1, float p1 = 1.70158);
